Revert "Add cancel, confirmation logic to placed request.(past and current)"
