+++ backend/TEMP_OTP.json
@@ -1,1 +1,1 @@
-{"dummy":"hi"}
+{"dummy":"hi","9704145656":{"type":"rider","otp":"190376","otpSetTime":1624259554967,"otpTriesLeft":5,"otpResendsLeft":0},"9999999999":{"type":"rider","otp":"444246","otpSetTime":1624259838198,"otpTriesLeft":5,"otpResendsLeft":5}}
