+++ frontend/src/components/global_ui/dialog/dialog.js
@@ -1,8 +1,5 @@
 import React from "react";
-import { useRef } from "react";
 import ReactDOM from "react-dom";
-import { useState } from "react/cjs/react.development";
-import { Spinner } from "../spinner";
 import "./dialog.css";
 /**
  * 
@@ -42,11 +39,15 @@ const Dialog = ({
       }
     : {};
 
+  
   return isShowing
     ? ReactDOM.createPortal(
         <React.Fragment>
           <div id="mw" className="modal-wrapper">
-            <div id="modal" className="modal">
+            <div
+              id="modal"
+              className="modal"
+            >
               <h3>{title}</h3>
               <p>{msg}</p>
 
@@ -54,12 +55,10 @@ const Dialog = ({
                 <button
                   className="modal-close-button"
                   onClick={() => {
-                    document
-                      .getElementById("mw")
-                      .classList.add("modal-wrapper-b");
+                    document.getElementById("mw").classList.add("modal-wrapper-b")
 
-                    document.getElementById("modal").classList.add("modal-b");
-                    setTimeout(() => onCancel(), 250);
+                  document.getElementById("modal").classList.add("modal-b")
+                  setTimeout(()=>onCancel(),250)
                   }}
                 >
                   Cancel
@@ -69,12 +68,11 @@ const Dialog = ({
                 style={style}
                 className="modal-close-button"
                 onClick={() => {
-                  document
-                    .getElementById("mw")
-                    .classList.add("modal-wrapper-b");
+                  document.getElementById("mw").classList.add("modal-wrapper-b")
 
-                  document.getElementById("modal").classList.add("modal-b");
-                  setTimeout(() => onOK(), 250);
+                  document.getElementById("modal").classList.add("modal-b")
+                  setTimeout(()=>onOK(),250)
+                  
                 }}
               >
                 Okay!
@@ -86,111 +84,5 @@ const Dialog = ({
       )
     : null;
 };
-/**
- *
- * Same as Dialog, but except you need to pass setDialogData and routeRedirect (takes you to that path after confirming)
- * EG:
- *      const [dialogData,setDialogData] = useState({show:false,msg:""})
- *
- * Refer Placed_requests.js for better understanding
- * @returns
- */
-const ConfirmDialog = ({
-  isShowing,
-  msg,
-  setDialogData,
-  title = "Alert",
-  onOK = async () => {},
-}) => {
-  const [loading, setLoading] = useState(false);
-  const onOKCompleted = useRef(false);
-  const [makeAlert, setMakeAlert] = useState(false);
-  const style = makeAlert
-    ? {
-        gridColumnStart: 1,
-        gridColumnEnd: -1,
-      }
-    : {};
 
-  return isShowing
-    ? ReactDOM.createPortal(
-        <React.Fragment>
-          <div id="mw" className="modal-wrapper">
-            <div id="modal" className="modal">
-              <h3>{title}</h3>
-              {loading ? (
-                <span className="spinnerModal">
-                  <Spinner radius="2" />
-                </span>
-              ) : (
-                <p>{msg}</p>
-              )}
-
-              {!loading && (
-                <>
-                  {!makeAlert && (
-                    <button
-                      className="modal-close-button"
-                      onClick={() => {
-                        document
-                          .getElementById("mw")
-                          .classList.add("modal-wrapper-b");
-                        document
-                          .getElementById("modal")
-                          .classList.add("modal-b");
-
-                        setTimeout(
-                          () =>
-                            setDialogData({
-                              show: false,
-                              msg: "",
-                            }),
-                          250
-                        );
-                      }}
-                    >
-                      Cancel
-                    </button>
-                  )}
-                  <button
-                    style={style}
-                    className="modal-close-button"
-                    onClick={async () => {
-                      if (onOKCompleted.current) {
-                        document
-                          .getElementById("mw")
-                          .classList.add("modal-wrapper-b");
-                        document
-                          .getElementById("modal")
-                          .classList.add("modal-b");
-                       
-                        setTimeout(() => {
-                          setDialogData({
-                            show: false,
-                            msg: "",
-                          });
-                          setMakeAlert(false);
-                          onOKCompleted.current = false;
-                        }, 250);
-                      } else {
-                        setLoading(true);
-                        await onOK();
-                        onOKCompleted.current = true;
-                        setLoading(false);
-                        setMakeAlert(true);
-                      }
-                    }}
-                  >
-                    Okay!
-                  </button>
-                </>
-              )}
-            </div>
-          </div>
-        </React.Fragment>,
-        document.body
-      )
-    : null;
-};
-
-export { Dialog, ConfirmDialog };
+export default Dialog;
