+++ frontend/src/components/requester/my_requests/placed_request.module.css
@@ -24,12 +24,6 @@ p {
   box-shadow: 0px 0px 12px rgba(0, 0, 0, 0.137);
   margin-bottom: 0.5em;
 }
-
-.address a{
-  text-align: center;
-  color: blue;
-}
-
 .address span:first-child {
   font-weight: bold;
 }
@@ -61,12 +55,33 @@ p {
   display: flex;
   flex-direction: column;
 }
+.singleImage {
+  width: fit-content;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 0.3em;
+}
 
+.imagesContainer {
+  display: grid;
+  margin-top: 1em;
+  width: min(90vw, 550px);
+  gap: 0.5em;
+  grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
+}
+
+.singleImage img {
+  height: auto;
+  box-shadow: 0px 0px 12px rgba(0, 0, 0, 0.137);
+
+  border-radius: 8px;
+  width: clamp(160px, 50%, 200px);
+}
 
 .buttonsContainer {
   display: flex;
   justify-content: space-around;
-  margin: 1rem;
   width: min(100%, 350px);
 }
 .buttonsContainer button {
